Return stdin to processes (#3)
* Use File Descriptor 9 For While Loops

* Update Apple Tab With File Condition

* File Check Gnome Tab If Exists

+++ apple_tab.sh
@@ -1,33 +1,39 @@
-# #!/bin/sh
+#!/bin/sh
 
 # Enable jobs control.
 set -m
+
+# Open file descriptor 9 read and write.
+exec 9<> /dev/tty
 
 # https://stackoverflow.com/questions/2172352/in-bash-how-can-i-check-if-a-string-begins-with-some-value#answer-18558871
 applelinebeginswith() { case $2 in "$1"*) true;; *) false;; esac; }
 
 apple_trap() {
-  # When the Boss process kills the expire new tab processes we get a race condition.
-  # It does not alter the functionality and error output is swallowed by /dev/null
-  APPLE_LINE_NUMBER=0
+  if [ -f "$APPLE_RELATIVE_DIRECTORY/meow-pids-$APPLE_GROUP.txt" ]; then
+    APPLE_LINE_NUMBER=0
 
-  while read -r apple_pid; do
-    # If the process is still alive then kill it.
-    if ps -p $apple_pid > /dev/null; then
-      echo "$APPLE_LINE_NUMBER. Killing  Process: $apple_pid"
-      kill -$APPLE_KILL_SIGNAL $apple_pid
-    else
-      echo "$APPLE_LINE_NUMBER. Process Already Dead: $apple_pid"
-    fi
+    while read -r apple_pid; do
+      # If the process is still alive then kill it.
+      if ps -p $apple_pid > /dev/null; then
+        echo "$APPLE_LINE_NUMBER. Killing  Process: $apple_pid"
+        kill -$APPLE_KILL_SIGNAL $apple_pid
+      else
+        echo "$APPLE_LINE_NUMBER. Process Already Dead: $apple_pid"
+      fi
 
-    APPLE_LINE_NUMBER=$((APPLE_LINE_NUMBER + 1))
-  done < $APPLE_RELATIVE_DIRECTORY/meow-pids-$APPLE_GROUP.txt
+      APPLE_LINE_NUMBER=$((APPLE_LINE_NUMBER + 1))
+    done < $APPLE_RELATIVE_DIRECTORY/meow-pids-$APPLE_GROUP.txt
 
-  # Delete the meow-pids-N.txt file.
-  rm $APPLE_RELATIVE_DIRECTORY/meow-pids-$APPLE_GROUP.txt &>/dev/null
-  rmdir $APPLE_RELATIVE_DIRECTORY &>/dev/null
+    # Delete the meow-pids-N.txt file.
+    rm $APPLE_RELATIVE_DIRECTORY/meow-pids-$APPLE_GROUP.txt &>/dev/null
+    rmdir $APPLE_RELATIVE_DIRECTORY &>/dev/null
 
-  echo "Process cleanup done for meow-pids-${APPLE_GROUP}.txt" && exit
+    echo "Process cleanup done for meow-pids-${APPLE_GROUP}.txt"
+  fi
+
+  # Close file descriptor 9.
+  exec 9>&-
 }
 
 trap apple_trap EXIT
@@ -38,7 +44,7 @@ APPLE_RELATIVE_DIRECTORY=""
 APPLE_KILL_SIGNAL=0
 
 # dir,kill-signal,group-#,endure/expire,commands...
-while read apple_file_line_command; do
+while read apple_file_line_command <&9; do
   if [ "$APPLE_FILE_INDEX" = 0 ]; then
     APPLE_RELATIVE_DIRECTORY=$apple_file_line_command
   elif [ "$APPLE_FILE_INDEX" = 1 ]; then
@@ -55,7 +61,7 @@ while read apple_file_line_command; do
   fi
 
   APPLE_FILE_INDEX=$((APPLE_FILE_INDEX + 1))
-done <<EOT
+done 9<<EOT
   `echo "$@" | sed 's/<meow-c>/\n/g'`
 EOT
